fix: Move files to an unwatched location
When files were copied here it was trying to create the folders here triggering another event to run this script. This will now copy to another location.

Updated logs but its still not writing to them, likely a permission issue.

Had to add a sleep of 5 seconds because incron wouldnt let me know once the entire file had been moved to this location. Ideally it would.

Dont move the root path so added validation here.